Added image upload/download (storing at the disk). eshop-admin module is complete.

+++ eshop-db/src/main/java/com/mvnikitin/eshop/repositories/CategoryRepository.java
@@ -1,9 +1,9 @@
 package com.mvnikitin.eshop.repositories;
 
-import com.mvnikitin.eshop.entities.Category;
+import com.mvnikitin.eshop.model.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoryRepository  extends JpaRepository<Category, Long> {
+public interface CategoryRepository  extends JpaRepository<Category, Integer> {
 }
